requirements.txt
dnspython>=2.6,<3.0
python-whois>=0.9,<1.0
pytest>=8.0,<9.0
